Pulse Refactor: Channels (#3094)
* Make the first round of channel changes: pulse_channels.py code is moved into channels.py.

* Improve documentation

* fixup doc
