+++ CHANGELOG.md
@@ -32,6 +32,9 @@ The format is based on [Keep a Changelog].
 
 ### Deprecated
 
+-   The full import path `qiskit.pulse.channels.pulse_channels` is deprecated;
+    code is moved to `qiskit.pulse.channels.channels` instead. Users should
+    keep importing from `qiskit.pulse.channels` (\#3094).
 -   The gates `U` and `CX` are being deprecated in favor of `u3` and
     `cx`. (\#2380)
 -   The gate `u0` is being deprecated in favor of using multiple `id` gates
